Level 50
[Update] Health type (int) -> (float)
[Add] Big Boss Heals it self

+++ Assets/Scripts/EnemySystem/IDamageable.cs
@@ -2,6 +2,6 @@ namespace EnemySystem
 {
     public interface IDamageable
     {
-        void Damage(int damageAmount);
+        void Damage(float damageAmount);
     }
 }
